move mapbox extension examples to one folder
