+++ test/apps/mapbox-extension/README.md
@@ -1,0 +1,20 @@
+Integrations with the @deck.gl/mapbox extension
+
+# Usage
+
+```bash
+yarn
+yarn start
+# or
+yarn start-local
+```
+
+This directory contains four examples:
+
+| API | Flavor | Map Lib | URL/Folder |
+| --- | --- | --- | --- |
+| `MapboxOverlay` | pure-js | `mapbox-gl-js` | `/mapbox-overlay` |
+| `MapboxOverlay` | react | `maplibre-gl-js`, `react-map-gl` | `/mapbox-overlay-react` |
+| `MapboxLayers` | pure-js | `maplibre-gl-js` | `/mapbox-layers` |
+| `MapboxLayers` | react | `mapbox-gl-js`, `react-map-gl` | `/mapbox-layers-react` |
+
