+++ test/apps/mapbox-extension/index.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset='UTF-8' />
+    <title>@deck.gl/mapbox examples</title>
+    <style>
+      body {margin: 0; width: 100vw; height: 100vh; overflow: hidden;}
+    </style>
+  </head>
+  <body>
+    <ul>
+      <li><a href="/mapbox-overlay/">mapbox-overlay</a>: the pure-js example for the <code>MapboxOverlay</code> API.</li>
+      <li><a href="/mapbox-overlay-react/">mapbox-overlay-react</a>: the React example for the <code>MapboxOverlay</code> API.</li>
+      <li><a href="/mapbox-layers/">mapbox-layers</a>: the pure-js example for the <code>MapboxLayers</code> API.</li>
+      <li><a href="/mapbox-layers-react/">mapbox-layers-react</a> the React example for the <code>MapboxLayers</code> API.</li>
+    </ul>
+  </body>
+</html>
