+++ test/apps/mapbox-extension/layers.js
