+++ test/apps/mapbox-extension/mapbox-layers-react/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset='UTF-8' />
-    <title>deck.gl example</title>
+    <title>MapboxLayers react example</title>
     <style>
       body {margin: 0; width: 100vw; height: 100vh; overflow: hidden;}
     </style>
@@ -10,5 +10,5 @@
   </head>
   <body>
   </body>
-  <script type="module" src='app.jsx'></script>
+  <script type="module" src='mapbox-layers-react.jsx'></script>
 </html>
