+++ test/apps/mapbox-extension/mapbox-overlay-react/index.html
@@ -2,12 +2,12 @@
 <html>
   <head>
     <meta charset='UTF-8' />
-    <title>deck.gl example</title>
+    <title>MapboxOverlay react example</title>
     <style>
       body {margin: 0; width: 100vw; height: 100vh; overflow: hidden;}
     </style>
   </head>
   <body>
   </body>
-  <script type="module" src='app.jsx'></script>
+  <script type="module" src='mapbox-overlay-react.jsx'></script>
 </html>
