+++ test/apps/mapbox-extension/mapbox-overlay/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset='UTF-8' />
-    <title>deck.gl example</title>
+    <title>MapboxOverlay pure-js example</title>
     <style>
       #map {
         position: fixed;
@@ -12,9 +12,10 @@
         bottom: 0;
       }
     </style>
+    <link href='https://api.tiles.mapbox.com/mapbox-gl-js/v0.44.2/mapbox-gl.css' rel='stylesheet' />
   </head>
   <body>
     <div id="map"></div>
   </body>
-  <script type="module" src='app.js'></script>
+  <script type="module" src='mapbox-overlay.js'></script>
 </html>
