+++ test/apps/mapbox-extension/mapbox-overlay/mapbox-overlay.js
@@ -1,7 +1,7 @@
 import mapboxgl from 'mapbox-gl';
 import {MapboxOverlay} from '@deck.gl/mapbox';
 import {ScatterplotLayer, ArcLayer} from '@deck.gl/layers';
-import {mapboxBuildingLayer, deckPoiLayer, deckRouteLayer} from './layers';
+import {mapboxBuildingLayer, deckPoiLayer, deckRouteLayer} from '../layers';
 
 // Set your mapbox token here
 mapboxgl.accessToken = process.env.MapboxAccessToken; // eslint-disable-line
