+++ test/apps/mapbox-extension/package.json
@@ -5,6 +5,7 @@
   },
   "dependencies": {
     "deck.gl": "^8.4.0",
+    "maplibre-gl": "^2.4.0",
     "mapbox-gl": "^2.0.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
