--- test/apps/mapbox-layers-react/.gitignore
@@ -1,1 +1,0 @@
-mapbox-gl-dev.js
