--- test/apps/mapbox-layers-react/README.md
@@ -1,5 +1,0 @@
-React integration with MapboxLayer API
-
-```js
-yarn start-local
-```
