--- test/apps/mapbox-layers/.gitignore
@@ -1,1 +1,0 @@
-mapbox-gl-dev.js
