--- test/apps/mapbox-layers/README.md
@@ -1,5 +1,0 @@
-Pure-js integration with MapboxLayer API
-
-```js
-yarn start-local
-```
