--- test/apps/mapbox-overlay-react/README.md
@@ -1,5 +1,0 @@
-React integration with MapboxOverlay API
-
-```js
-yarn start-local
-```
