--- test/apps/mapbox-overlay-react/package.json
@@ -1,16 +1,0 @@
-{
-  "scripts": {
-    "start": "vite --open",
-    "start-local": "vite --config ../vite.config.local.mjs"
-  },
-  "dependencies": {
-    "deck.gl": "^8.4.0",
-    "maplibre-gl": "^2.4.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-map-gl": "^7.0.0"
-  },
-  "devDependencies": {
-    "vite": "^4.0.0"
-  }
-}
