--- test/apps/mapbox-overlay/.gitignore
@@ -1,1 +1,0 @@
-mapbox-gl-dev.js
