--- test/apps/mapbox-overlay/README.md
@@ -1,9 +1,0 @@
-Integration with MapboxOverlay API
-
-```js
-yarn start-local
-```
-
-This directory contains two independent examples:
-- `app.js`: the pure-js example entry point. URL: `/` (root)
-- `react-app.js`: the React example entry point. URL: `/reactApp`
